created factories for componenents
