+++ eunomiaJava/fitchTextEditor/src/main/java/nl/rug/editorFrame/ComponentFactory.java
@@ -1,0 +1,6 @@
+package nl.rug.editorFrame;
+
+public interface ComponentFactory {
+
+    public void createJComponent(int token);
+}
